Add `Calendar` and `Popover` UI components, enhance `Follower` API with filtering and admin-only access, and update dependencies

+++ src/server/routers/_app.ts
@@ -35,9 +35,10 @@ export const appRouter = router({
 		
 		return {
 			user,
-			accessToken: sign({sub: user.username}, 'qwerty123')
+			accessToken: sign({id: user.id, username: user.username, role: user.role.name}, 'qwerty123')
 		}
 	}),
+	
 	getTotalUsers: procedure.input(z.object({
 		role: z.string().optional()
 	})).query(async ({input}) => {
@@ -60,7 +61,14 @@ export const appRouter = router({
 			orderBy: z.enum(['id', 'fullname', 'username', 'roleId', 'createdAt']).optional().default('createdAt'),
 			orderSort: z.enum(['asc', 'desc']).optional().default('desc')
 		}))
-		.query(async ({input}) => {
+		.query(async ({input, ctx}) => {
+			if (!ctx.user || ctx.user.role !== 'Administrator') {
+				throw new TRPCError({
+					message: 'Not Allowed',
+					code: 'UNAUTHORIZED'
+				})
+			}
+			
 			const where: Prisma.UserWhereInput = {}
 			if (input.role) {
 				where.role = {
@@ -84,7 +92,14 @@ export const appRouter = router({
 		username: z.string(),
 		password: z.string(),
 		roleId: z.number()
-	})).mutation(async ({input}) => {
+	})).mutation(async ({input, ctx}) => {
+		if (!ctx.user || ctx.user.role !== 'Administrator') {
+			throw new TRPCError({
+				message: 'Not Allowed',
+				code: 'UNAUTHORIZED'
+			})
+		}
+		
 		const {password, ...rest} = input
 		return prisma.user.create({
 			data: {
@@ -99,7 +114,14 @@ export const appRouter = router({
 		fullname: z.string(),
 		username: z.string(),
 		password: z.string().optional()
-	})).mutation(async ({input}) => {
+	})).mutation(async ({input, ctx}) => {
+		if (!ctx.user || ctx.user.role !== 'Administrator') {
+			throw new TRPCError({
+				message: 'Not Allowed',
+				code: 'UNAUTHORIZED'
+			})
+		}
+		
 		const {id, password, ...rest} = input
 		
 		const data: Prisma.UserUpdateInput = rest
@@ -116,7 +138,14 @@ export const appRouter = router({
 	
 	deleteUser: procedure.input(z.object({
 		id: z.number()
-	})).mutation(async ({input}) => {
+	})).mutation(async ({input, ctx}) => {
+		if (!ctx.user || ctx.user.role !== 'Administrator') {
+			throw new TRPCError({
+				message: 'Not Allowed',
+				code: 'UNAUTHORIZED'
+			})
+		}
+		
 		const {id} = input
 		
 		return prisma.user.delete({
@@ -124,60 +153,103 @@ export const appRouter = router({
 		})
 	}),
 	
-	getTotalFollowers: procedure.query(async ({}) => {
-		return prisma.follower.count()
-	}),
-	
-	getFollowers: procedure.input(z.object({
-		page: z.number().optional().default(1),
-		search: z.string().optional(),
-		memberId: z.number().optional().nullish()
+	getTotalFollowers: procedure.input(z.object({
+		hasMember: z.boolean().optional().default(false)
 	})).query(async ({input}) => {
-		const where: Prisma.FollowerWhereInput = {}
-		
-		if (input.search) {
-			where.OR = [
-				{
-					member: {
-						fullname: {
-							contains: `%${input.search}%`
-						}
-					}
-				},
-				{
-					fullname: {
-						contains: `%${input.search}%`
-					}
-				}
-			]
+		const where: Prisma.FollowerWhereInput = {
+			memberId: null
 		}
 		
-		if (input.memberId) {
-			where.memberId = input.memberId
-		}
-		
-		const offset = input.page - 1
-		const take = 10
-		return prisma.follower.findMany({
-			skip: offset * take,
-			take,
-			where,
-			include: {
-				member: true
+		if (input.hasMember) {
+			where.memberId = {
+				not: null
 			}
+		}
+		
+		return prisma.follower.count({
+			where
 		})
 	}),
+	
+	getFollowers: procedure
+		.input(z.object({
+			page: z.number().optional().default(1),
+			search: z.string().optional(),
+			memberId: z.number().optional().nullish(),
+			hasMember: z.boolean().optional().default(false)
+		}))
+		.query(async ({input, ctx}) => {
+			if (!ctx.user) {
+				throw new TRPCError({
+					message: 'Not Allowed',
+					code: 'UNAUTHORIZED'
+				})
+			}
+			
+			const where: Prisma.FollowerWhereInput = {}
+			
+			if (input.search) {
+				where.OR = [
+					{
+						member: {
+							fullname: {
+								contains: `%${input.search}%`,
+								mode: 'insensitive'
+							}
+						}
+					},
+					{
+						fullname: {
+							contains: `%${input.search}%`,
+							mode: 'insensitive'
+						}
+					}
+				]
+			}
+			
+			if (input.memberId) {
+				where.memberId = input.memberId
+			} else if (input.hasMember) {
+				where.memberId = {
+					not: null
+				}
+			} else if (ctx.user.role === 'Administrator') {
+				where.memberId = null
+			} else {
+				where.memberId = ctx.user.id
+			}
+			
+			const offset = input.page - 1
+			const take = 10
+			return prisma.follower.findMany({
+				skip: offset * take,
+				take,
+				where,
+				include: {
+					member: true
+				}
+			})
+		}),
 	
 	createFollower: procedure.input(z.object({
 		fullname: z.string(),
 		phoneNumber: z.string(),
 		arrivalDate: z.string()
-	})).mutation(async ({input}) => {
+	})).mutation(async ({input, ctx}) => {
+		if (!ctx.user) {
+			return prisma.follower.create({
+				data: {
+					...input,
+					arrivalDate: new Date(input.arrivalDate)
+				}
+			})
+		}
+		
 		return prisma.follower.create({
 			data: {
 				...input,
 				arrivalDate: new Date(input.arrivalDate),
-				memberId: 2
+				memberId: ctx.user.id
 			}
 		})
 	}),
@@ -187,7 +259,14 @@ export const appRouter = router({
 		fullname: z.string(),
 		phoneNumber: z.string(),
 		arrivalDate: z.string()
-	})).mutation(async ({input}) => {
+	})).mutation(async ({input, ctx}) => {
+		if (!ctx.user) {
+			throw new TRPCError({
+				message: 'Not Allowed',
+				code: 'UNAUTHORIZED'
+			})
+		}
+		
 		const {id, arrivalDate, ...rest} = input
 		return prisma.follower.update({
 			where: {
@@ -196,18 +275,26 @@ export const appRouter = router({
 			data: {
 				...rest,
 				arrivalDate: new Date(arrivalDate),
-				memberId: 2
+				memberId: ctx.user.id
 			}
 		})
 	}),
 	
 	deleteFollower: procedure.input(z.object({
 		id: z.number()
-	})).mutation(async ({input}) => {
+	})).mutation(async ({input, ctx}) => {
+		if (!ctx.user) {
+			throw new TRPCError({
+				message: 'Not Allowed',
+				code: 'UNAUTHORIZED'
+			})
+		}
+		
 		const {id} = input
 		return prisma.follower.delete({
 			where: {
-				id
+				id,
+				memberId: ctx.user.id
 			}
 		})
 	})
